feat: add tiles in a reasonable way
What a day. I started by designing a few tiles in pyxel, but then
when I tried to import them the import failed. For the longest time
I thought it must be pyxel adding some extra data to the png that
rgbgfx was not liking. With some help from Valentina on the discord
I stumbled upon the real answer: I was using the number of bytes
in the file to loop over the data, and that number quickly gets to
be larger than 8 bits :D amazing

Once that was sorted out I did a bit of tweaking to the way meta
tiles are copied to VRAM so that my meta tile definitions can just
be straight 1, 2, 3, 4 instead of 1, 3, 2, 4

Then I added some nice grass tiles to the overworld to test... and
everything was fine.